coreset prototype

+++ coreset/utils.py
@@ -107,7 +107,7 @@ def random_sampler(data, K):
 
 @timeit
 def craig_baseline(data, K):
-    features = data.astype(np.float32)
+    features = data.astype(np.float16)
     D = pairwise_distances(features, metric="euclidean", n_jobs=20)
     D = D.max() - D
     V = np.arange(len(features), dtype=int).reshape(-1, 1)
